Add glue to hook up APZ machinery.
This adds some glue code to allow creating the APZ machinery. With this patch
the machinery is still disabled, but it can be enabled by setting the pref
apz.allow_without_e10s to true. If that pref is set, APZ will handle input
events, but the scrolling will still be synchronous. That is, the visible scroll
positions will only update on a new main thread repaint/layer transaction, and
not asynchronously. That will come later. The experience with the pref enabled
is much worse because it causes the main thread to skip a lot of paints on the
assumption that APZ will do async scrolling, which it doesn't yet. Also various
clipping/overflow issues manifest with this pref on which makes the experience
quite bad.