Pass bounds and overflow to wr_pop_dl_builder
We're going to need to have the ability to pass both the bounds and overflow
rects from Gecko Layers to WR stacking contexts, so this updates the function
signature to allow that. Also uses WRRect instead of x/y/width/height. No
functional changes; all the values are the same as they were before.