Push a new stacking context for each scrollframe.
When rendering a layer, ensure that we push a new stacking context for
each scroll metadata on that layer. Right now these stacking contexts
don't have scroll ids but they will in the next patch.

Note that each stacking context also changes the coordinate space, so
some adjustments need to be made to the existing coordinates being
passed around, so that they are relative to the most recent enclosing
stacking context, which is either the parent ContainerLayer's visible
region, or the nearest ancestor scrollable FrameMetrics' composition
bounds.