Fix up coordinates when creating WR items.
This updates the coordinates passed to WR display items and stacking contexts
so that they are more in line with what WR is expecting. Each stacking context
that is created provides reasonable bounds and overflow rect values based on
the visible region, relative to the parent stacking context. Similarly display
items provide coordinates relative to the nearest stacking context.

Note that as part of this change we stop passing the layer transform to the
WR stacking contexts. This is because of a semantic difference between how
transforms are used in the gecko layer tree and how they are used in the WR
stacking contexts. In particular, the gecko layer tree provides a way to clip
layer contents after they have gone through the transform. However the WR
stacking contexts apply their transform after the clip to the bounds of the
stacking context. This mismatch means that it might be better to just incorporate
the transform into the layer properties on the Gecko side, which is what this
patch does. It will probably break down when the transforms include scaling
factors.

Also this patch removes some unneccessary stacking context creation in some of
the layer types.

Positioning of color layers is still a little broken (it was broken before this
patch as well), and will need some further investigation.