Reduce difference between {WebRender,Client}LayerManager.
This patch updates some bits of code that assume a ClientLayerManager to
not assume that. It also updates WebRenderLayerManager to behave a little
bit more like a ClientLayerManager. Particular in terms of creating it
and setting it up; it can now be done without requiring a special method
in nsBaseWidget to create the WebRenderLayerManager.