Make e10s work on Windows.
When we try to create a GLContext on windows, we initialize the WGLLibrary
on the compositor thread, which means we try to use gfxInfo on the compositor
thread. (Because we need to check if FEATURE_DX_INTEROP2 is blacklisted).
This patch is a hacky workaround to prevent that blowing up.

dvander says a more correct solution is probably to store the blacklist status
in gfxConfig somewhere during startup, so it's readily available to the compositor
thread.