Update uses of #ifdef MOZ_ENABLE_WEBRENDER.
We *have* to include PWebRenderBridge.ipdl both with and without WR, because
IPDL doesn't support #ifdefing. This leads to a situation where if we try to
ifdef out parts of the C++ code we end up with a lot of #ifdefs everywhere.
Instead, I chose to have almost all of the C++ code build regardless of whether
WR is enabled. However, the creation of the WRLayerManager vs ClientLayerManager
is guarded by an ifdef, and the webrender.h file has stub functions if we not
building with WR. This still ensures that we don't build any Rust WR code unless
WR is enabled.